lanchain-openai-examples/api-key/requirements.txt
langchain-openai
python-dotenv